plots: try to address chart height challenge
The label for the last / max tick on the
y axis may have its upper bound _above_
that very last tick, making the plot
a few pixels larger in height compared to
a plot that does not have a tick at the
upper bound of the y axis. This is
tracked here:

https://github.com/vega/vega/issues/2860

I have tried using the vega tick parameters

    labelFlush / labelFlushOffset

for the y axis and they indeed can be used
to change the position of the upper bound
tick label. However, this also changes the
position of the 0 (lower bound) tick label
and that looks weird.

I have tried playing with the vega tick
label parameter labelBound=True via
Altair's

    .configure_axisY(labelBound=True)

and that works (it hides the upper tick label)
but _also_ hides the zero tick label.

All of this is now summarized in

https://github.com/vega/vega/issues/2860#issuecomment-751511501