report: add toc, add headings
"--from=markdown_strict+pandoc_title_block"
was used to allow for adding raw HTML
to the markdown document.

This however makes it so that a the auto-
generated table of contents (--toc) does
not have links to the sections.

Use pandoc markdown instead of markdown_strict
and allow for adding html via the native_divs
extension. as of pandoc documentation <script>
tags are always included verbatim.