Fix exit on errexit when return code is used for checks
At the beginning of the script, the return code of `git ls-remote`
is used in a check while `errexit` is active. This does not work
because in case the `git ls-remote` returns a non-zero code, the
script will exit immediately and the line checking the return code
in an `if` clause is never reached.

For this to work, `errexit` needs to be temporarily disabled and
the return code stored in a variable for later use.