style: referrer/path legend: top, instead of right
Add color coding legend for the top referrer/path
plots above the plots.

That means: make the plots as wide as all the other
plots, and also equally wide among the two of them.

That allows for showing the same time range across
all plots with the same relationship between
plot width and time width across all plots. I wanted
to do this for a long time, for better interpretability.