analyze.py: gen_date_axis_lim(): allow for zero length DFs
Allow for some items in the `dfs` iterable to
be of zero length.

Require that at least one item is of non-zero
length.

Allowing for zero length df items is the direct fix
for the traceback shown in #41.

A more global fix for #41 is to properly deal with
empty stargazer/fork time series